+++ tests/formatter/input-output-pairs/multiline-properties.out.gd
@@ -3,10 +3,8 @@ var p1:
 var p2:
 	set = __set,
 	get = __get
-
 var p3:
 	get = __get
-
 var p4:
 	get = __get,
 	set = __set
@@ -16,6 +14,12 @@ var p5:
 		pass
 var p6:
 	get:
+		pass
+var p7:
+	set(_x):
+		pass
+var p8:
+	set(_x):
 		pass